requirements.txt
Flask==3.0.0
Pillow==10.1.0
numpy==1.26.2

